+++ ElectricalProgressive-Core/Utils/PathCacheManager.cs
@@ -13,7 +13,7 @@ namespace ElectricalProgressive.Utils
         private class Entry
         {
             public BlockPos[]? Path;
-            public int[]? FacingFrom;
+            public byte[]? FacingFrom;
             public bool[][]? NowProcessedFaces;
             public Facing[]? UsedConnections;
             public DateTime LastAccessed;
@@ -61,9 +61,10 @@ namespace ElectricalProgressive.Utils
         /// Попытаться получить путь из кэша.
         /// </summary>
         public static bool TryGet(
-            BlockPos start, BlockPos end,
+            BlockPos start,
+            BlockPos end,
             out BlockPos[] path,
-            out int[] facingFrom,
+            out byte[] facingFrom,
             out bool[][] nowProcessed,
             out Facing[] usedConnections,
             out int version)
@@ -92,9 +93,11 @@ namespace ElectricalProgressive.Utils
         /// Сохранить в кэше новый путь или обновить существующий.
         /// </summary>
         public static void AddOrUpdate(
-            BlockPos start, BlockPos end, int currentVersion,
+            BlockPos start,
+            BlockPos end,
+            int currentVersion,
             BlockPos[] path,
-            int[] facingFrom,
+            byte[] facingFrom,
             bool[][] nowProcessedFaces,
             Facing[] usedConnections)
         {
